@@ -3,9 +3,9 @@ package com.livelike.engagementsdk.chat
 import android.annotation.SuppressLint
 import android.content.Context
 import android.os.Build
-import android.support.v13.view.inputmethod.EditorInfoCompat
-import android.support.v13.view.inputmethod.InputConnectionCompat
-import android.support.v7.widget.AppCompatEditText
+import androidx.core.view.inputmethod.EditorInfoCompat
+import androidx.core.view.inputmethod.InputConnectionCompat
+import androidx.appcompat.widget.AppCompatEditText
 import android.util.AttributeSet
 import android.view.MotionEvent
 import android.view.View
