@@ -1,7 +1,7 @@
 package com.livelike.engagementsdk.chat
 
 import android.graphics.Rect
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 
 class VerticalSpaceItemDecoration(private val verticalSpaceHeight: Int) :
