@@ -2,13 +2,13 @@
 
 package com.livelike.engagementsdk.chat.stickerKeyboard
 
-import android.support.annotation.Px
-import android.support.v7.recyclerview.extensions.ListAdapter
-import android.support.v7.util.DiffUtil
-import android.support.v7.widget.GridLayoutManager
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.PagerSnapHelper
-import android.support.v7.widget.RecyclerView
+import androidx.annotation.Px
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.PagerSnapHelper
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -68,7 +68,8 @@ class StickerCollectionViewHolder(
 ) : RecyclerView.ViewHolder(itemView) {
 
     init {
-        itemView.rvStickers.layoutManager = GridLayoutManager(itemView.context, 6)
+        itemView.rvStickers.layoutManager =
+            GridLayoutManager(itemView.context, 6)
         itemView.empty_recent_text.setTextColor(emptyRecentTextColor)
     }
 
