@@ -1,7 +1,7 @@
 package com.livelike.engagementsdk.widget
 
 import android.content.Context
-import android.support.constraint.ConstraintLayout
+import androidx.constraintlayout.widget.ConstraintLayout
 import android.util.AttributeSet
 import android.view.View
 import com.livelike.engagementsdk.AnalyticsService
