@@ -3,8 +3,8 @@ import android.animation.Animator
 import android.animation.AnimatorSet
 import android.animation.ObjectAnimator
 import android.content.Context
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.widget.FrameLayout
 import com.livelike.engagementsdk.EngagementSDK
@@ -54,7 +54,8 @@ class WidgetsTimeLineView(
                 sdk
             )
         adapter.list.addAll(timeLineViewModel.timeLineWidgets)
-        timeline_rv.layoutManager = LinearLayoutManager(context)
+        timeline_rv.layoutManager =
+            LinearLayoutManager(context)
         timeline_rv.adapter = adapter
         initListeners()
     }
