@@ -1,7 +1,7 @@
 package com.livelike.engagementsdk.widget.view
 
 import android.content.Context
-import android.support.constraint.ConstraintLayout
+import androidx.constraintlayout.widget.ConstraintLayout
 import android.util.AttributeSet
 import android.widget.FrameLayout
 import com.livelike.engagementsdk.MockAnalyticsService
