@@ -2,7 +2,7 @@ package com.livelike.engagementsdk.widget.view.components
 
 import android.animation.Animator
 import android.content.Context
-import android.support.constraint.ConstraintLayout
+import androidx.constraintlayout.widget.ConstraintLayout
 import android.util.AttributeSet
 import android.view.View
 import com.livelike.engagementsdk.DismissAction
