@@ -1,7 +1,7 @@
 package com.livelike.livelikedemo
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.widget.TextView
 import com.livelike.engagementsdk.LiveLikeContentSession
 import com.livelike.engagementsdk.core.data.models.LeaderBoard
