@@ -1,9 +1,9 @@
 package com.livelike.livelikedemo
 
 import android.os.Bundle
-import android.support.design.widget.TabLayout
-import android.support.v4.view.ViewPager
-import android.support.v7.app.AppCompatActivity
+import com.google.android.material.tabs.TabLayout
+import androidx.viewpager.widget.ViewPager
+import androidx.appcompat.app.AppCompatActivity
 import com.livelike.livelikedemo.channel.ChannelManager
 import com.livelike.livelikedemo.ui.main.SectionsPagerAdapter
 
