@@ -1,7 +1,7 @@
 package com.livelike.livelikedemo
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 
 class WidgetStandaloneActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
