@@ -1,7 +1,7 @@
 package com.livelike.livelikedemo.customwidgets
 
 import android.content.Context
-import android.support.constraint.ConstraintLayout
+import androidx.constraintlayout.widget.ConstraintLayout
 import android.util.AttributeSet
 import android.view.View
 import com.bumptech.glide.Glide
