@@ -2,9 +2,9 @@ package com.livelike.livelikedemo.customwidgets
 
 import android.content.Context
 import android.graphics.Color
-import android.support.constraint.ConstraintLayout
-import android.support.v7.widget.GridLayoutManager
-import android.support.v7.widget.RecyclerView
+import androidx.constraintlayout.widget.ConstraintLayout
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import android.util.AttributeSet
 import android.view.LayoutInflater
 import android.view.View
@@ -50,7 +50,11 @@ class CustomPollWidget : ConstraintLayout {
         pollWidgetModel?.widgetData?.let { liveLikeWidget ->
             liveLikeWidget.options?.let {
                 if (it.size > 2) {
-                    rcyl_poll_list.layoutManager = GridLayoutManager(context, 2)
+                    rcyl_poll_list.layoutManager =
+                        GridLayoutManager(
+                            context,
+                            2
+                        )
                 }
                 val adapter =
                     PollListAdapter(context, isImage, ArrayList(it.map { item -> item!! }))
