@@ -2,8 +2,8 @@ package com.livelike.livelikedemo.customwidgets
 
 import android.content.Context
 import android.os.Handler
-import android.support.constraint.ConstraintLayout
-import android.support.v7.widget.GridLayoutManager
+import androidx.constraintlayout.widget.ConstraintLayout
+import androidx.recyclerview.widget.GridLayoutManager
 import android.util.AttributeSet
 import android.view.View
 import com.livelike.engagementsdk.widget.widgetModel.FollowUpWidgetViewModel
@@ -13,7 +13,8 @@ import kotlinx.android.synthetic.main.custom_poll_widget.view.button2
 import kotlinx.android.synthetic.main.custom_poll_widget.view.imageView2
 import kotlinx.android.synthetic.main.custom_poll_widget.view.rcyl_poll_list
 
-class CustomPredictionWidget : ConstraintLayout {
+class CustomPredictionWidget :
+    ConstraintLayout {
     var predictionWidgetViewModel: PredictionWidgetViewModel? = null
     var followUpWidgetViewModel: FollowUpWidgetViewModel? = null
     var isImage = false
@@ -51,7 +52,11 @@ class CustomPredictionWidget : ConstraintLayout {
         widgetData?.let { liveLikeWidget ->
             liveLikeWidget.options?.let {
                 if (it.size > 2) {
-                    rcyl_poll_list.layoutManager = GridLayoutManager(context, 2)
+                    rcyl_poll_list.layoutManager =
+                        GridLayoutManager(
+                            context,
+                            2
+                        )
                 }
                 val adapter =
                     PollListAdapter(context, isImage, ArrayList(it.map { item -> item!! }))
