@@ -3,8 +3,8 @@ import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.support.constraint.ConstraintLayout
-import android.support.v4.content.ContextCompat
+import androidx.constraintlayout.widget.ConstraintLayout
+import androidx.core.content.ContextCompat
 import android.view.View
 import com.bumptech.glide.Glide
 import com.livelike.engagementsdk.widget.viewModel.WidgetStates
