@@ -1,6 +1,6 @@
 import android.app.Application
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 
 import com.livelike.livelikedemo.LiveLikeApplication
 
